Substitui antigo script de óbitos por novo (registral)
