+++ schema/obito_cartorio.csv
@@ -1,15 +1,35 @@
 field_name,field_type
 date,date
 state,text
-new_deaths_pneumonia_2019,integer
-new_deaths_pneumonia_2020,integer
-new_deaths_covid19,integer
-new_deaths_respiratory_failure_2019,integer
-new_deaths_respiratory_failure_2020,integer
 epidemiological_week_2019,integer
 epidemiological_week_2020,integer
-deaths_covid19,integer
-deaths_respiratory_failure_2019,integer
-deaths_respiratory_failure_2020,integer
+new_deaths_sars_2019,integer
+new_deaths_pneumonia_2019,integer
+new_deaths_respiratory_failure_2019,integer
+new_deaths_septicemia_2019,integer
+new_deaths_indeterminate_2019,integer
+new_deaths_others_2019,integer
+new_deaths_sars_2020,integer
+new_deaths_pneumonia_2020,integer
+new_deaths_respiratory_failure_2020,integer
+new_deaths_septicemia_2020,integer
+new_deaths_indeterminate_2020,integer
+new_deaths_others_2020,integer
+new_deaths_covid19,integer
+deaths_sars_2019,integer
 deaths_pneumonia_2019,integer
+deaths_respiratory_failure_2019,integer
+deaths_septicemia_2019,integer
+deaths_indeterminate_2019,integer
+deaths_others_2019,integer
+deaths_sars_2020,integer
 deaths_pneumonia_2020,integer
+deaths_respiratory_failure_2020,integer
+deaths_septicemia_2020,integer
+deaths_indeterminate_2020,integer
+deaths_others_2020,integer
+deaths_covid19,integer
+new_deaths_total_2019,integer
+new_deaths_total_2020,integer
+deaths_total_2019,integer
+deaths_total_2020,integer
